Rewrite pattern for expand-stick
