+++ src/Accelerators/NNPA/Conversion/ONNXToZHigh/ONNXToZHighCommon.hpp
@@ -113,12 +113,6 @@ mlir::ValueRange splitAlongAxis(
     onnx_mlir::MultiDialectBuilder<onnx_mlir::OnnxBuilder> &create,
     mlir::Value X, int64_t axis);
 
-// Check if a value is a constant tensor of a single f32 value or not.
-bool isF32ScalarConstantTensor(mlir::Value v);
-
-// Get FloatAttr from a constant tensor of a single f32 value.
-mlir::FloatAttr getScalarF32AttrFromConstant(mlir::Value v);
-
 // Emit ONNX Concat to store the shape of the input x.
 mlir::Value getDynShape(
     mlir::Location loc, mlir::PatternRewriter &rewriter, mlir::Value x);
